# docs/openapi/keyai-api-v1.yaml
openapi: 3.0.3
info:
  title: KeyAI Microservices API
  version: 1.0.0
  description: |
    Complete API documentation for KeyAI's microservices architecture.
    
    ## Services Overview
    
    - **Auth Service** (Port 3005): Authentication and user management
    - **Capture Service** (Port 3001): Keyboard event capture
    - **Masker Service** (Port 3002): PII detection and masking
    - **Storage Service** (Port 3003): Data persistence
    - **Search Service** (Port 3004): Hybrid search capabilities
    
  contact:
    name: KeyAI Team
    email: api@keyai.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Kong API Gateway (Production)
  - url: http://localhost:{port}
    description: Direct service access (Development)
    variables:
      port:
        enum:
          - "3001"
          - "3002"
          - "3003"
          - "3004"
          - "3005"
        default: "3001"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /capture/start:
    post:
      tags: [capture]
      summary: Start keyboard capture
      operationId: startCapture
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Capture started
        '400':
          description: Capture already active
  /capture/stop:
    post:
      tags: [capture]
      summary: Stop keyboard capture
      operationId: stopCapture
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Capture stopped
        '400':
          description: Capture not active
  
  # Masker Service Endpoints
...